continuous training SP fully implemented, not varified

+++ MARL/SP_train.py
@@ -12,7 +12,7 @@ from examples.rllib import utils
 from meltingpot.python import substrate
 
 def get_config(
-    substrate_name: str = "chicken_in_the_matrix__repeated_medium",
+    substrate_name: str = "chicken_in_the_matrix__repeated_obstacle",
     num_rollout_workers: int = 2,
     rollout_fragment_length: int = 100,
     train_batch_size: int = 1600,
@@ -124,40 +124,63 @@ def main():
   tune.register_env("meltingpot", utils.env_creator)
 
   # parameters
-  save_path = '/home/yuxin/meltingpot/MARL/SP_logs/c5m'
-  checkpoints_path = '/home/yuxin/meltingpot/MARL/SP_checkpoints/c5m'
-  log_path = '/home/yuxin/meltingpot/MARL/SP_outputs/c5m.txt'
-  checkpoint_freq = 5
-  num_gens = 5
+  save_path = '/home/yuxin/meltingpot/MARL/SP_logs/c5o'
+  checkpoints_path = '/home/yuxin/meltingpot/MARL/SP_checkpoints/c5o'
+  log_path = '/home/yuxin/meltingpot/MARL/SP_outputs/c5o.txt'
+  checkpoint_freq = 125
+  num_gens = 25
   seeds = [11, 22, 33, 44, 55]
-  starting_timestep = 0
+
+  continuous_training = True
+  starting_gen = 11
+  starting_timestep = (starting_gen+1) * checkpoint_freq * config.train_batch_size
 
   gen_len = checkpoint_freq * config.train_batch_size
   num_seeds = len(seeds)
 
   # clear output
-  f = open(log_path, "w")
-  f.close()
+  if not continuous_training:
+    f = open(log_path, "w")
+    f.close()
 
   # Initialize ray, train and save
   ray.init()
 
   # logging
-  timesteps = [[] for j in range(num_seeds)]
-  policy_reward_min = [[[] for i in range(2)] for j in range(num_seeds)]
-  policy_reward_mean = [[[] for i in range(2)] for j in range(num_seeds)]
-  policy_reward_max = [[[] for i in range(2)] for j in range(num_seeds)]
+  if not continuous_training:
+    timesteps = [[] for j in range(num_seeds)]
+    policy_reward_min = [[[] for i in range(2)] for j in range(num_seeds)]
+    policy_reward_mean = [[[] for i in range(2)] for j in range(num_seeds)]
+    policy_reward_max = [[[] for i in range(2)] for j in range(num_seeds)]
+  else:
+    raw_data = np.load(save_path+'.npz')
+    timesteps = raw_data['timesteps'].tolist()
+    policy_reward_min = raw_data['policy_reward_min'].tolist()
+    policy_reward_mean = raw_data['policy_reward_mean'].tolist()
+    policy_reward_max = raw_data['policy_reward_max'].tolist()
 
   # A dictionary of lists of checkpoints for each seed/population
 
-  for f in os.scandir(checkpoints_path):
-    if f.is_dir():
-      shutil.rmtree(f)
-  checkpoints_dict = {}
-  for seed in range(num_seeds):
-    checkpoints_dict[f'Seed_{seed}'] = []
+  if not continuous_training:
+    for f in os.scandir(checkpoints_path):
+      if f.is_dir():
+        shutil.rmtree(f)
+    checkpoints_dict = {}
+    for seed in range(num_seeds):
+      checkpoints_dict[f'Seed_{seed}'] = []
+  else:
+    checkpoints_dict = {}
+    for seed in range(num_seeds):
+      current_seed_path = os.path.join(checkpoints_path, f'seed_{seed}')
+      dir_list = sorted(os.listdir(current_seed_path))
+      latest_gen_path = os.path.join(current_seed_path, dir_list[-1])
+      latest_checkpoint_path = os.path.join(latest_gen_path, os.listdir(latest_gen_path)[-1])
+      checkpoints_dict[f'Seed_{seed}'] = [latest_checkpoint_path]
 
   for gen in range(num_gens):
+    if continuous_training:
+      if gen <= starting_gen:
+        continue
     for seed in range(num_seeds):
       # config.env_config["seed"] = seeds[seed]
       ppo = PPO(config=config.to_dict())
@@ -201,7 +224,10 @@ def main():
           f.writelines(lines)
         f.close()
         # save results
-        timesteps[seed].append(gen * gen_len + results["timesteps_total"]+starting_timestep)
+        if continuous_training:
+          timesteps[seed].append(gen * gen_len + results["timesteps_total"]+starting_timestep)
+        else:
+          timesteps[seed].append(gen * gen_len + results["timesteps_total"])
         policy_reward_min[seed][0].append(results["policy_reward_min"]["agent_0"] if results["policy_reward_min"] else float('NaN'))
         policy_reward_min[seed][1].append(results["policy_reward_min"]["agent_1"] if results["policy_reward_min"] else float('NaN'))
         policy_reward_mean[seed][0].append(results["policy_reward_mean"]["agent_0"] if results["policy_reward_mean"] else float('NaN'))
